+++ strawberry/aiohttp/views.py
@@ -20,10 +20,10 @@ from strawberry.aiohttp.handlers import (
     GraphQLTransportWSHandler,
     GraphQLWSHandler,
 )
-from strawberry.http.base_view import (
-    AsyncBaseHTTPView,
+from strawberry.http.async_base_view import AsyncBaseHTTPView
+from strawberry.http.exceptions import HTTPException
+from strawberry.http.typevars import (
     Context,
-    HTTPException,
     RootValue,
 )
 from strawberry.subscriptions import GRAPHQL_TRANSPORT_WS_PROTOCOL, GRAPHQL_WS_PROTOCOL
