+++ strawberry/chalice/views.py
@@ -3,7 +3,12 @@ from __future__ import annotations
 from typing import TYPE_CHECKING, Any, Dict, List, Mapping, Optional, Union
 
 from chalice.app import Request, Response
-from strawberry.http.base_view import BaseHTTPView, Context, HTTPException, RootValue
+from strawberry.http.sync_base_view import (
+    BaseSyncHTTPView,
+    Context,
+    HTTPException,
+    RootValue,
+)
 from strawberry.http.temporal_response import TemporalResponse
 from strawberry.utils.graphiql import get_graphiql_html
 
@@ -45,7 +50,7 @@ class ChaliceHTTPRequestAdapter:
         return self.request.headers.get("Content-Type", None)
 
 
-class GraphQLView(BaseHTTPView[Request, Response, Context, RootValue]):
+class GraphQLView(BaseSyncHTTPView[Request, Response, Context, RootValue]):
     allow_queries_via_get: bool = True
     request_adapter_class = ChaliceHTTPRequestAdapter
