+++ strawberry/fastapi/router.py
@@ -34,10 +34,10 @@ from strawberry.fastapi.handlers import GraphQLTransportWSHandler, GraphQLWSHand
 from strawberry.http import (
     process_result,
 )
-from strawberry.http.base_view import (
-    AsyncBaseHTTPView,
+from strawberry.http.async_base_view import AsyncBaseHTTPView
+from strawberry.http.exceptions import HTTPException
+from strawberry.http.typevars import (
     Context,
-    HTTPException,
     RootValue,
 )
 from strawberry.subscriptions import GRAPHQL_TRANSPORT_WS_PROTOCOL, GRAPHQL_WS_PROTOCOL
