+++ strawberry/http/exceptions.py
@@ -1,0 +1,4 @@
+class HTTPException(Exception):
+    def __init__(self, status_code: int, reason: str):
+        self.status_code = status_code
+        self.reason = reason
