+++ strawberry/http/typevars.py
@@ -1,0 +1,6 @@
+from typing import TypeVar
+
+Request = TypeVar("Request", contravariant=True)
+Response = TypeVar("Response")
+Context = TypeVar("Context")
+RootValue = TypeVar("RootValue")
